# GUID lookup optimizations and benchmarks