--- a/src/WinRT.Runtime/Interop/IContextCallback.cs
+++ b/src/WinRT.Runtime/Interop/IContextCallback.cs
@@ -39,6 +39,8 @@ namespace ABI.WinRT.Interop
     [Guid("000001da-0000-0000-C000-000000000046")]
     internal sealed unsafe class IContextCallback : global::WinRT.Interop.IContextCallback
     {
+        internal static readonly Guid IID = new(0x000001da, 0, 0, 0xC0, 0, 0, 0, 0, 0, 0, 0x46);
+
         [Guid("000001da-0000-0000-C000-000000000046")]
         public struct Vftbl
         {