--- a/src/WinRT.Runtime/Interop/IMarshal.cs
+++ b/src/WinRT.Runtime/Interop/IMarshal.cs
@@ -132,7 +132,7 @@ namespace ABI.WinRT.Interop
             {
                 EnsureHasFreeThreadedMarshaler();
 
-                Guid iid_IUnknown = typeof(IUnknownVftbl).GUID;
+                Guid iid_IUnknown = IUnknownVftbl.IID;
                 Guid iid_unmarshalClass;
                 t_freeThreadedMarshaler.GetUnmarshalClass(&iid_IUnknown, IntPtr.Zero, MSHCTX.InProc, IntPtr.Zero, MSHLFLAGS.Normal, &iid_unmarshalClass);
                 return iid_unmarshalClass;