--- a/src/WinRT.Runtime/Interop/IWeakReferenceSource.net5.cs
+++ b/src/WinRT.Runtime/Interop/IWeakReferenceSource.net5.cs
@@ -49,6 +49,16 @@ namespace WinRT.Interop
 
             return ComWrappersSupport.CreateCCWForObject<IUnknownVftbl>(target, riid);
         }
+
+        public IntPtr ResolveForABI(Guid riid)
+        {
+            if (!_ref.TryGetTarget(out object target))
+            {
+                return IntPtr.Zero;
+            }
+
+            return ComWrappersSupport.CreateCCWForObjectForABI(target, riid);
+        }
     }
 }
 
@@ -99,9 +109,7 @@ namespace ABI.WinRT.Interop
 
             try
             {
-                using var objRef = ComWrappersSupport.CreateCCWForObject(new global::WinRT.Interop.ManagedWeakReference(ComWrappersSupport.FindObject<object>(thisPtr)));
-                ExceptionHelpers.ThrowExceptionForHR(objRef.TryAs(IWeakReference.IID, out var ptr));
-                *weakReference = ptr;
+                *weakReference = ComWrappersSupport.CreateCCWForObjectForABI(new global::WinRT.Interop.ManagedWeakReference(ComWrappersSupport.FindObject<object>(thisPtr)), IWeakReference.IID);
             }
             catch (Exception __exception__)
             {
@@ -134,14 +142,20 @@ namespace ABI.WinRT.Interop
         [UnmanagedCallersOnly(CallConvs = new[] { typeof(CallConvStdcall) })]
         private static int Do_Abi_Resolve(IntPtr thisPtr, Guid* riid, IntPtr* objectReference)
         {
-            IObjectReference _objectReference = default;
-
             *objectReference = default;
 
             try
             {
-                _objectReference = global::WinRT.ComWrappersSupport.FindObject<global::WinRT.Interop.IWeakReference>(thisPtr).Resolve(*riid);
-                *objectReference = _objectReference?.GetRef() ?? IntPtr.Zero;
+                var weakReference = global::WinRT.ComWrappersSupport.FindObject<global::WinRT.Interop.IWeakReference>(thisPtr);
+                if (weakReference is global::WinRT.Interop.ManagedWeakReference managedWeakReference)
+                {
+                    *objectReference = managedWeakReference.ResolveForABI(*riid);
+                }
+                else
+                {
+                    using var _objectReference = weakReference.Resolve(*riid);
+                    *objectReference = _objectReference?.GetRef() ?? IntPtr.Zero;
+                }
             }
             catch (Exception __exception__)
             {