--- a/src/WinRT.Runtime/Marshalers.cs
+++ b/src/WinRT.Runtime/Marshalers.cs
@@ -1011,8 +1011,7 @@ namespace WinRT
     {
         private static readonly Type HelperType = typeof(T).GetHelperType();
         private static Func<T, IObjectReference> _ToAbi;
-        private static Func<IntPtr, T> _FromAbi;
-        private static Func<IObjectReference, IObjectReference> _As;
+        private static Func<T, IObjectReference> _CreateMarshaler;
 
         public static T FromAbi(IntPtr ptr)
         {
@@ -1046,14 +1045,12 @@ namespace WinRT
                 return ObjectReference<IUnknownVftbl>.Attach(ref ptr);
             }
 
-            if (_As is null)
+            if (_CreateMarshaler is null)
             {
-                _As = BindAs();
+                _CreateMarshaler = BindCreateMarshaler();
             }
 
-            var inspectable = MarshalInspectable<T>.CreateMarshaler(value, true);
-
-            return _As(inspectable);
+            return _CreateMarshaler(value);
         }
 
         public static IntPtr GetAbi(IObjectReference value) => 
@@ -1065,17 +1062,14 @@ namespace WinRT
 
         public static IntPtr FromManaged(T value)
         {
-            if (value is null)
-            {
-                return IntPtr.Zero;
-            }
-            return CreateMarshaler(value).GetRef();
+            using var objRef = CreateMarshaler(value);
+            return objRef?.GetRef() ?? IntPtr.Zero;
         }
 
         public static unsafe void CopyManaged(T value, IntPtr dest)
         {
-            *(IntPtr*)dest.ToPointer() =
-                (value is null) ? IntPtr.Zero : CreateMarshaler(value).GetRef();
+            using var objRef = CreateMarshaler(value);
+            *(IntPtr*)dest.ToPointer() = objRef?.GetRef() ?? IntPtr.Zero;
         }
 
         public static unsafe MarshalInterfaceHelper<T>.MarshalerArray CreateMarshalerArray(T[] array) => MarshalInterfaceHelper<T>.CreateMarshalerArray(array, (o) => CreateMarshaler(o));
@@ -1094,16 +1088,6 @@ namespace WinRT
 
         public static unsafe void DisposeAbiArray(object box) => MarshalInterfaceHelper<T>.DisposeAbiArray(box);
 
-        private static Func<IntPtr, T> BindFromAbi()
-        {
-            var fromAbiMethod = HelperType.GetMethod("FromAbi", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Static);
-            var objReferenceConstructor = HelperType.GetConstructor(BindingFlags.NonPublic | BindingFlags.CreateInstance | BindingFlags.Instance, null, new[] { fromAbiMethod.ReturnType }, null);
-            var parms = new[] { Expression.Parameter(typeof(IntPtr), "arg") };
-            return Expression.Lambda<Func<IntPtr, T>>(
-                    Expression.New(objReferenceConstructor,
-                        Expression.Call(fromAbiMethod, parms[0])), parms).Compile();
-        }
-
         private static Func<T, IObjectReference> BindToAbi()
         {
             var parms = new[] { Expression.Parameter(typeof(T), "arg") };
@@ -1113,22 +1097,20 @@ namespace WinRT
                     HelperType.GetField("_obj", BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.DeclaredOnly)), parms).Compile();
         }
 
-        private static Func<IObjectReference, IObjectReference> BindAs()
+        private static Func<T, IObjectReference> BindCreateMarshaler()
         {
             var vftblType = HelperType.FindVftblType();
+            Guid iid = GuidGenerator.GetIID(HelperType);
             if (vftblType is not null)
             {
-                var parms = new[] { Expression.Parameter(typeof(IObjectReference), "arg") };
-                return Expression.Lambda<Func<IObjectReference, IObjectReference>>(
-                    Expression.Call(
-                        parms[0],
-                        typeof(IObjectReference).GetMethod("As", Type.EmptyTypes).MakeGenericMethod(vftblType)
-                        ), parms).Compile();
+                var methodInfo = typeof(MarshalInspectable<T>).GetMethod("CreateMarshaler", new Type[] { typeof(T), typeof(Guid), typeof(bool) }).
+                    MakeGenericMethod(vftblType);
+                var createMarshaler = (Func<T, Guid, bool, IObjectReference>) methodInfo.CreateDelegate(typeof(Func<T, Guid, bool, IObjectReference>));
+                return obj => createMarshaler(obj, iid, true);
             }
             else
             {
-                Guid iid = GuidGenerator.GetIID(HelperType);
-                return obj => obj.As<IUnknownVftbl>(iid);
+                return obj => MarshalInspectable<T>.CreateMarshaler<IUnknownVftbl>(obj, iid, true);
             }
         }
     }