--- a/src/WinRT.Runtime/Projections/ICustomPropertyProvider.net5.cs
+++ b/src/WinRT.Runtime/Projections/ICustomPropertyProvider.net5.cs
@@ -285,6 +285,8 @@ namespace ABI.Microsoft.UI.Xaml.Data
         private static readonly ManagedCustomPropertyProviderVftbl AbiToProjectionVftable;
         public static readonly IntPtr AbiToProjectionVftablePtr;
 
+        internal static readonly Guid IID = new(0x7C925755, 0x3E48, 0x42B4, 0x86, 0x77, 0x76, 0x37, 0x22, 0x67, 0x03, 0x3F);
+
         static unsafe ManagedCustomPropertyProviderVftbl()
         {
             AbiToProjectionVftable = new ManagedCustomPropertyProviderVftbl