--- a/src/WinRT.Runtime/Projections/ICustomPropertyProvider.netstandard2.0.cs
+++ b/src/WinRT.Runtime/Projections/ICustomPropertyProvider.netstandard2.0.cs
@@ -273,6 +273,8 @@ namespace ABI.Microsoft.UI.Xaml.Data
 
         private static readonly Delegate[] DelegateCache = new Delegate[4];
 
+        internal static readonly Guid IID = new(0x7C925755, 0x3E48, 0x42B4, 0x86, 0x77, 0x76, 0x37, 0x22, 0x67, 0x03, 0x3F);
+
         static unsafe ManagedCustomPropertyProviderVftbl()
         {
             AbiToProjectionVftable = new ManagedCustomPropertyProviderVftbl