--- a/src/cswinrt/strings/WinRT.cs
+++ b/src/cswinrt/strings/WinRT.cs
@@ -846,8 +846,12 @@ namespace WinRT
             }
         }
     }
-}
 
+    internal static class InterfaceIIDs
+    {
+        internal static readonly Guid IInspectable_IID = new(0xAF86E2E0, 0xB12D, 0x4c6a, 0x9C, 0x5A, 0xD7, 0xAA, 0x65, 0x10, 0x1E, 0x90);
+    }
+}
 
 namespace System.Runtime.CompilerServices
 {