… src/Benchmarks/GuidPerf.cs
… src/Benchmarks/QueryInterface.cs