- Moved a couple more GUIDs to hard coded values to reduce custom attribute lookup and added some caches
- Moved to using IsDefined rather than GetCustomAttribute where possible as benchmarks showed it being more performant and didn't do an allocation
- Added benchmarks for GUID lookup
- Improved performance of weak references by adding a new GetCCW API which avoids the IObjectReference allocation.  A future change being experimented with might make use of this concept in all FromManaged scenarios where today we allocate an IObjectReference when we can avoid it. 
- Reduced an IObjectReference allocation in CreateMarshaler to do the QI separately by rather having MarshalInspectable do it via the overloaded method.

Before PR

|                                    Method |       Mean |    Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|------------------------------------------ |-----------:|---------:|----------:|-------:|-------:|----------:|
|           GetWeakReferenceOfManagedObject | 1,947.4 ns | 38.28 ns |  67.04 ns | 0.0610 |      - |     264 B |
| GetAndResolveWeakReferenceOfManagedObject | 2,422.6 ns | 48.17 ns | 101.60 ns | 0.0839 | 0.0420 |     360 B |
|            GetWeakReferenceOfNativeObject |   292.6 ns |  2.59 ns |   2.16 ns | 0.0057 |      - |      24 B |


PR

|                                    Method |       Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------------------------------------ |-----------:|---------:|---------:|-------:|----------:|
|           GetWeakReferenceOfManagedObject | 1,772.6 ns | 16.49 ns | 13.77 ns | 0.0401 |     168 B |
| GetAndResolveWeakReferenceOfManagedObject | 1,934.1 ns | 14.65 ns | 12.99 ns | 0.0381 |     168 B |
|            GetWeakReferenceOfNativeObject |   286.1 ns |  1.05 ns |  0.88 ns | 0.0057 |      24 B |

Before PR

|                            Method |        Mean |     Error |    StdDev |  Gen 0 | Allocated |
|---------------------------------- |------------:|----------:|----------:|-------:|----------:|
|                      GetClassGuid |  1,309.6 ns |   9.91 ns |   9.27 ns | 0.0076 |      32 B |
|                   GetDelegateGuid |    334.5 ns |   1.29 ns |   1.14 ns | 0.0076 |      32 B |
|                    CreateListGuid | 21,081.1 ns | 146.41 ns | 136.95 ns | 0.8240 |   3,531 B |
| CreateDictionaryWithStringKeyGuid | 22,383.6 ns |  88.66 ns |  74.04 ns | 0.9460 |   4,036 B |
|   CreateDictionaryWithBoolKeyGuid | 22,899.6 ns | 371.13 ns | 397.10 ns | 0.9460 |   3,996 B |
|        CreateReadOnlyEnumListGuid | 16,040.9 ns |  69.55 ns |  58.08 ns | 0.9460 |   4,042 B |
|    CreateReadOnlyFlagEnumListGuid | 16,844.8 ns |  75.75 ns |  67.15 ns | 1.0376 |   4,450 B |
|      CreateReadOnlyStructListGuid |  8,807.9 ns | 101.64 ns |  79.35 ns | 0.4730 |   2,041 B |
|   CreateReadOnlyInterfaceListGuid | 13,245.7 ns |  61.14 ns |  57.19 ns | 0.5341 |   2,265 B |
|       CreateReadOnlyClassListGuid | 20,869.7 ns | 106.96 ns |  94.82 ns | 0.7935 |   3,347 B |
|    CreateReadOnlyDelegateListGuid | 13,455.0 ns |  88.94 ns |  69.44 ns | 0.5341 |   2,297 B |

PR

|                            Method |       Mean |    Error |    StdDev |     Median |  Gen 0 | Allocated |
|---------------------------------- |-----------:|---------:|----------:|-----------:|-------:|----------:|
|                      GetClassGuid | 1,282.3 ns |  4.09 ns |   3.62 ns | 1,281.9 ns | 0.0076 |      32 B |
|                   GetDelegateGuid |   331.5 ns |  0.96 ns |   0.75 ns |   331.5 ns | 0.0076 |      32 B |
|                    CreateListGuid | 2,641.5 ns | 12.19 ns |  10.80 ns | 2,640.7 ns | 0.4654 |   1,961 B |
| CreateDictionaryWithStringKeyGuid | 2,824.0 ns | 13.22 ns |  12.37 ns | 2,821.1 ns | 0.5302 |   2,233 B |
|   CreateDictionaryWithBoolKeyGuid | 2,917.5 ns | 55.12 ns | 126.65 ns | 2,863.9 ns | 0.5226 |   2,193 B |
|        CreateReadOnlyEnumListGuid | 5,008.4 ns | 54.33 ns |  48.16 ns | 5,002.4 ns | 0.4196 |   1,784 B |
|    CreateReadOnlyFlagEnumListGuid | 2,712.6 ns | 38.35 ns |  32.02 ns | 2,709.1 ns | 0.3548 |   1,496 B |
|      CreateReadOnlyStructListGuid | 2,392.6 ns | 16.58 ns |  14.70 ns | 2,389.8 ns | 0.3738 |   1,569 B |
|   CreateReadOnlyInterfaceListGuid | 2,311.8 ns | 13.71 ns |  12.82 ns | 2,311.6 ns | 0.3319 |   1,392 B |
|       CreateReadOnlyClassListGuid | 2,649.4 ns | 11.49 ns |   9.60 ns | 2,651.9 ns | 0.4425 |   1,865 B |
|    CreateReadOnlyDelegateListGuid | 2,347.9 ns | 22.73 ns |  17.74 ns | 2,340.7 ns | 0.3510 |   1,472 B |